+++ src/CarRental.Application/Features/Vehicles/Commands/CreateVehicle/CreateVehicleCommand.cs
@@ -20,7 +20,7 @@ public record CreateVehicleCommand : IRequest<Result<VehicleDto>>
     /// <summary>Gets or sets the PlateNumber.</summary>
     public string PlateNumber { get; init; } = string.Empty;
     /// <summary>Gets or sets the CurrentMileage.</summary>
-    public float CurrentMileage { get; init; }
+    public decimal CurrentMileage { get; init; }
     /// <summary>Gets or sets the ClassificationId.</summary>
     public long ClassificationId { get; init; }
     /// <summary>Gets or sets the Transmission.</summary>
